model_training/requirements.txt
pandas
google-cloud-bigquery[bqstorage,pandas]
google-cloud-storage
pystan==2.19.1.1
fbprophet